make all pages protected
